Implementen alignment and closing for Linux

+++ multi_window_linux/lib/multi_window_linux.dart
@@ -42,8 +42,8 @@ class MultiWindowLinux extends MultiWindowInterface {
       'title': title,
       'alignment': alignment != null
           ? {
-              'x': alignment.x,
-              'y': alignment.y,
+              'x': alignment.x.toInt(),
+              'y': alignment.y.toInt(),
             }
           : null,
     });
